feat({score.ly,parts.ly}:score.layout): use letters for rehearsal marks
Measure numbers are to combersome for both commenting within music.ly
(add a measure? go tick up every single measure number comment in your
piece) as well as difficult to line up to the new use of section
variables in music.ly:global (sectionASpace, etc.) and for chord
changes. Letter rehearsal marks absorb changes better and are easier to
track accross the code and variables in the source files.