fix({parts,score}.ly): move score/part line break helper functions
`lyInclude/definitions.ily` contains defaults so our songs always compile.
But the redefining of these vars needs to come after the importing of
lyInclude and before the importing of the music to take effect. So we
moved where they were in the files from the Settings section to the
Include section.