feat({parts,score}.ly): include settings.ily, add globalTranspose()
A `settings.ily` file is now included in the music repository that can
include helper functions that act as a type of global setting.

globalTranspose():

Is a wrapper function that the score and all
individual parts can be run through that can transpose everything as a
final step before being compiled into score pdfs. All individual
instrument transpositions and key changes within the piece are
maintained.

Care must be made to specific selection of the values of globalTranspose
so that proper enharmonics are selected. But at least there is one
single place in which to set this.