feat({parts,score}.ly): add scoreStyle()
`scoreStyle()` is defined in the music repository in the new
settings.ily file. It too is a wrapper function that will use tags to
filter out that which is tagged "concert" or "transposed" in the
layout.ily file thereby selecting music that is kept in concert pitch
(for for reading the score entirely in concert picth, and also
potentially allowing concert pitch midi playback) while still allowing
individual parts, or the score itself, to display transposed parts if
desired.

The staves, chords, and global are joined into a single variable which
then gets passed twice to the parts (ie. "partA"). Filtering results in
removal of what would be redundant music; the entire piece of music
twice in a single score, but one in concert and one transposed.