feat({parts,score}.ly): add concert pitch midi playback
Midi files used to be subject to the same transposition/key as the pdf
output. So now it's split into it's own score and uses only the concert
pitch music via tags for midi output. The parts and score can display
concert or transposed pitches now independent of the midi output, which
is always concert pitch.

The midi is still subject to the globalTranspose() setting.